Tests

+++ src/core/cpuCO2.py
@@ -3,6 +3,7 @@ from PyQt5.QtWidgets import QApplication
 import sys
 from core.gui import PlotWindowApp
 
+
 def cpu_co2(country_code = None) -> None:
     """Function to run the core of the app: compare CPU usage with CO2 emissions"""
 
@@ -55,7 +56,8 @@ def cpu_co2(country_code = None) -> None:
               'database, it is therefore not possible to retrieve CO2 emissions data.')
 
 
-def combine_cpu_CO2(cpu_usage: float, usage_time: int, cpu_tdp: int, co2_intensity: float) -> float:
+def combine_cpu_CO2(cpu_usage: float, usage_time: float, 
+                    cpu_tdp: float, co2_intensity: int) -> float:
     """
     Combine CPU usage, CPU TDP and CO2 intensity to compute CO2 consumption.
 
@@ -70,6 +72,7 @@ def combine_cpu_CO2(cpu_usage: float, usage_time: int, cpu_tdp: int, co2_intensi
     :return: grams of CO2 emitted for your CPU usage in the specified time
     :rtype: float
     """
+
     # transfrom Watt in kWatt
     cpu_tdp_KW = cpu_tdp/1000
 
@@ -82,7 +85,8 @@ def combine_cpu_CO2(cpu_usage: float, usage_time: int, cpu_tdp: int, co2_intensi
     return co2_emissions
 
 
-def get_interval_emissions(cpu_tdp: int, co2_intensity: float, time_frequency: int) -> float:
+def get_interval_emissions(cpu_tdp: float, co2_intensity: int, 
+                           time_frequency: int) -> float:
     """
     From a given cpu_tdp and co2_intensity return the co2 emissions computing the cpu_usage at current time.
